+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format located [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- metric-zfs-arc.rb, aggregate ZFS ARC metrics (@phumpal)
 
 ## [2.0.1] - 2018-02-27
 ### Changed